move AWS/GCP/HTTP error functions to respective files
