+++ pkg/fetcher/fetcher_http.go
@@ -242,3 +242,16 @@ func (r *httpResource) Length(ctx context.Context) (int64, *ResourceError) {
 	}
 	return size, nil
 }
+
+func HTTPStatusToException(status int) *ResourceError {
+	if status == 0 {
+		return nil
+	}
+
+	switch status {
+	case http.StatusOK, http.StatusCreated, http.StatusAccepted, http.StatusPartialContent, http.StatusNoContent, http.StatusResetContent, http.StatusNotModified:
+		return nil
+	default:
+		return NewResourceError(ResourceErrorCode(status))
+	}
+}
